Refill with fuel added
had to make some changes, Item now contains the enum itemType.

added a function to search by itemType in the inventory, went for the easy route of using the getcomponent fuelmanager and the bool 'refillfuel'.

then remove the item with the inventory.removeitem.

inventory.removeitem created a null item instead of a blank item like the rest so i changed that too.

finally rerouted the outOfFuel event to the cutengine of root player2 (there are 2 on the scene for some reason)